+++ routes/web.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\ThreadController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\HelpController;
 use App\Http\Controllers\PostController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
@@ -23,7 +24,11 @@ Route::get('/contact', function () {
 
 Route::get('/dashboard', function () {
     return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+})->middleware(['auth', 'verified'])
+    ->name('dashboard');
+
+Route::patch('/dashboard/update/{user}', [ProfileController::class, 'updateFor'])->middleware(['auth', 'verified'])
+    ->name('dashboard.update');
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])
@@ -33,6 +38,9 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])
         ->name('profile.destroy');
 });
+
+Route::get('/users', [UserController::class, 'index'])
+    ->name('users.index');
 
 Route::get('/discussion-board', function () {
     return view('discussion-board');